+++ src/p_ceilng.c
@@ -76,7 +76,7 @@ void T_MoveCeiling(ceiling_t *ceiling)
             // UP
             res = T_MovePlane(ceiling->sector, ceiling->speed, ceiling->topheight, false, 1, ceiling->direction);
 
-            if (!(leveltime & 7))
+            if (!(maptime & 7))
                 switch (ceiling->type)
                 {
                     case silentCrushAndRaise:
@@ -131,7 +131,7 @@ void T_MoveCeiling(ceiling_t *ceiling)
             // DOWN
             res = T_MovePlane(ceiling->sector, ceiling->speed, ceiling->bottomheight, ceiling->crush, 1, ceiling->direction);
 
-            if (!(leveltime & 7))
+            if (!(maptime & 7))
                 switch (ceiling->type)
                 {
                     case silentCrushAndRaise: